Upgrade to LLVM 21 (#31)
`cargo test` passes at least. Although I also updated `setup.sh`, it
looks like [`llvm@21` may not exist in Homebrew yet][1]. @femtomc's
changes in mlir-rs/mlir-sys#71 suggest an alternative way to install
LLVM in CI.

Still, this works end-to-end with Melior for my use cases when combined
with mlir-rs/mlir-sys#71 and mlir-rs/melior#731, if that is useful to
know.

I don't know if this warrants a version bump as
https://github.com/mlir-rs/tblgen-rs/pull/25 did (seems like the
maintainer's call)

(Succeeds #28 which I made from the wrong branch)

[1]: https://formulae.brew.sh/formula/